It's ALIVE!!!11
But seriously though, this is pretty neat.

Enemy boids (teapots) try to attack the friendly (ball) boids but are scared off by the big bad monkey (player). The teapots make jabs at the friendlies like hunters poking at a school of fish. If enemies get too close to the flock, the flock members try to leap away. If the flock is abandoned, the teapots go in for the kill whilst the friendly tries to hopelessly run away.